Start the butchering
