+++ windows-gaming/Cargo.toml
@@ -12,4 +12,3 @@ time = "0.1.37"
 clap = "2.26"
 common = { path = "common" }
 driver = { path = "driver" }
-wizard = { path = "wizard" }
