+++ windows-gaming/config/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wizard"
+name = "config"
 version = "0.1.0"
 authors = ["main() <main@ehvag.de>"]
